Squashed audio engine changes
Audio engine updates are driven from Kotlin by the orientation callback.
That's the main event that affcects the 3D audio and so it seems like a
sensible point in time to update the audio positioning.

Calling createBeacon in Kotlin creates an audio beacon of the type set
by setBeaconType at the location provided.
createTextToSpeech outputs the speech provided from the text string at
the location provided.