Add mock google-services.json for use in actions
When running the run-test action to test a pull request, the user
may not have repo write access. This means that they won't have
access to the GitHub secrets. This mock google-services.json ensures
that the build will still work. It's over-written by the 'secret'
file when building a release build.