Tidy up lint warnings and land some old work (#20)
* Ensure FMOD_RESULT is defined to aid code inspection

fmod_codec.h was using FMOD_RESULT without a definition and that was
leading to various lint warnings. Adding the include of fmod_common.h
resolves this.

* Clear up TextToSpeech progress listener onError warning

UtteranceProgressListener::OnError is deprecated, but we still need
to define it for now. Adding the @Deprecated message stops this from
being a lint warning.

* Load FMOD at earliest point

We want the FMOD library to be loaded before the app starts calling
initialization functions on it. Loading it with the application should
do this.

* Tidy up application behaviour

Two changes here:

1. Clicking to stop the foreground service now exits the application.
2. The MainActivity launchMode is now singleTask.

Combined, these two gives a slightly more obvious behaviour when
starting the application and running it via intents.

* Extend notification to open app

Clicking on the notification now opens the app making it easier to stop
the audio.