Extend Route database for geospatial support
This commit alters the RoutePoint class so that it can be used with the
Realm geospatial functions. Whilst doing this it was noticed that the
lifetime of waypoints wasn't being correctly handled. Deleting a
RouteData in realm won't delete the child RoutePoints and so we do that
by hand. After deleting a RouteData any RoutePoint with empty backlinks
to routes are also deleted.