Fix GpxTest tests to all run standalone without dependency on each other
The GpxTest tests were reliant on the order in which they were run. The
parsing tests were populating the inMemory database and then the database
tests were testing that. However, when Firebase runs the test the database
from previous tests doesn't exist and so the tests were failing. This
change makes each of the tests standalone without dependency on other tests.