Add getListOfBeaconTypes to AudioEngine and instrumentation test for it
getListOfBeaconTypes returns a list of the names of the available beacons within the
AudioEngine. This is to allow the UI to interrogate the AudioEngine rather than having
two separate lists of beacon types. In addition, I didn't understand enough about
Instrumentation tests when I wrote the AudioEngine code. Now that I do, I've added a
belated set of initial tests.