Various AudioEngine tests and improvements
AudioEngine.setBeaconType now takes a String which is the name of the
Beacon type to use. The test loops to test every Beacon type that
AudioEngine reports, rotating the listener so that the sound for each
of the orientation angles is played. This is done fairly quickly so
as not to make the test over long.

Mentions of com_scottishtecharmy_soundscape were replaced with the correct
org_scottishtecharmy_soundscape.

Added awaitTextToSpeechInitialization which blocks all AudioEngine speech
related APIs until the TextToSpeech engine has fully initialized. Without
this there's a high risk of speech being dropped.

Added APIs to get the available voices and languages.