Creation of OnboardingActivity
There are a number of changes in this commit, and I'm really just
submitting it for discussion. I'm not wholly clear that these are
the right approaches!

I've moved the onboarding screens into their own OnboardingActivity.
The initial reason for this was because I didn't like the behaviour
of the back action in the app. During onboarding, it moved back
through the screensm but once at the Home screen it would still move
back to the Finish screen. By moving the onboarding screens into their
own Activity I found that it was easier to discard the action history
that got the user to that point. Now the behaviour is that going back
from the home page exits the app - which is the same whether the
onboarding was run on that session or not. Activities seem to be "old
fashioned" and so perhaps this isn't the right approach, but it does
keep all of that code in it's own little world which seems neater.

The other change I made within the Screens was to pass in an onNavigate
callback rather than a NavHostController. That seems slighly neater and
was recommended in the Android docs as it keeps the navigation stuff
out of the UI. The next changes will involve ViewModel injection, and I
was just trying to match the code to the docs as much as possible.

This commit also includes the permissions messing around. However,
because the permissions are currently in MainActivity, none of those are
currently requested in the onboarding screens. Once we have some
ViewModels in place, then the permissions can be moved to one of those
and the requests should happen at the right point in time.