Initial audio support for onboarding screens
This commit adds ViewModels for the Hearing, Language and AudioBeacons Screens.
It adds them within the same files as the UIs because they are small and it's
easier to see what's going on. The NativeAudioEngine is now a singleton and
is automatically created on first use.
We want @Preview to carry on working on all of the streams, so for those the
ViewModels are left uninitiated and mock data is supplied by the @Preview
functions. This seems a reasonable way of fixing it up, though I'm less clear
if it's the most correct way.