Update to screenshot test approach
This change adds a number of different configurations for which the
Preview is run:

 * Locales - all of the languages that we support
 * Devices - tablet, small phone and medium phone
 * Fonts - smaller and larger

Currently this results in 119 files. The names change each time the
update builds is run. It doesn't make sense to pollute our main
git branch with these, so I've added them to a .gitignore.
For github actions, the approach I'll likely take is:

 * First new action which runs the update screenshot build and then stores
   the resulting screenshots in a reference artifact for the repo
 * Second action runs the screenshot validation against the retrieved
   reference artifact. Ideall it would then publish the resuling
   org.scottishtecharmy.soundscape.PreviewTest.html to the GitHub pages
   site.
 * Third action runs the update screenshot build but discards the results

The last of these is the most useful in that it would hopfully be useful
for checking whenever Previews get broken which otherwise go unnoticed in
the build. The first action could be run manually to get a quick overview of
the current state of the UI. The second might be more useful once we've got
a more stable application.

The main problem with not storing the reference images in the repo is that
rewinding to an older commit won't rewind the reference and so building on a
branch would get confusing. Needs some more thinking about.