Try including screenshots in documentation compilation
We could include the report from the preview screenshot validation test,
but that requires validating against reference images which we don't have
yet. Instead, we build the reference images and use some very basic HTML
to display all of the screenshots for each Test. The full resolution test
can be saved from the link and so this should give us automatic screenshots
of each screen in all locale and a couple of screen sizes.