Deliver onboarding screens' ViewModel data as flows
Firebase was reporting a test issue:

Your app uses 1 non-SDK interfaces, which are incompatible with Android P+.
   Ljava/lang/invoke/MethodHandles$Lookup;-><init>(Ljava/lang/Class;I)V

It turns out that initializing ViewModel data inside an init function is a
bad idea - see
  https://www.droidcon.com/2024/03/22/mastering-android-viewmodels-essential-dos-and-donts-part-1-%f0%9f%9b%a0%ef%b8%8f/

This change removes the init from AudioBeaconsViewModel and instead uses a
flow to initialize the list of beacon types. That initialization will now
happen when the UI code calls into the ViewModel with
collectAsStateWithLifecycle instead of when theAudioBeaconsViewModel was
created.

The LanguageViewModel has been changed in the same way. The list of supported
languages is now in a flow and is only initialized once.