Remove screenshots from docs but build them everywhere else
Putting the screenshots in the docs increased the overhead of docs
compilation by a factor of 10 and the result wasn't that readable.
This change removes that and instead we build screenshots during
all of the other actions. Release and nightly builds both upload
them as artifacts which should make them easily accessible for
developeres.