Reduce minSdk version and increase targetSdk version
minSdk can drop to 29 because it turns out that the splash screen
is supported on that release too, just not as well. The splash.xml
has to be versioned due to the less good support.
AndroidStudio has been complaining about the targetSdk not being the
latest, so bump that up at the same time. In doing so move the
dependencies to using libs.versions.toml which is the current way to
handle package versions.