Move location and orientation out into separate class
This is the start of adding alternative location and orientation providers.
The location and orientation providers have been moved into their own class
and the option of creating a static location provider has been added. That
teleports the user to a fixed location but allows the orientation still to
be controlled.