Upgrade libraries including maplibre
Upgrading MapLibre removed all of the outdated MapBox naming so now
it's all MapLibre which makes more sense.
Almost all of the warnings in libs.versions.toml were resolved which
included:

 * Upgrading libraries
 * Removing duplicate declarations
 * Removing declarations of two different versions of a library (okHttp)

The only remaining out of date library is Realm. If we bump that to 2.x
then it doesn't compile (known issues it seems). Otherwise, everything
seems to be working the same as before.