Fix initial audio beacon buffer direction
The first buffer played for an audio beacon was always the 'straight
ahead' one. This change ensures that on creation the beacon uses the
current listener location to select the correct initial audio
direction. This is easy to demonstrate in testing - simply long click
on the map in different directions from the listener and immediately
hear the possible sounds that the current beacon type can make.