Replace button that stops service and exits with sleep toggle
The iOS app had a 'sleep' button in the top right and this change
moves towards that model. Clicking on the button toggles the running
of the foreground service. When the service isn't running, there's no
location/heading update and no text to speech. The icon showing the
current location on the map is removed, and then when the service is
restarted it reappears and the map is recentered.

Not sure what iOS does in this case, need to compare.