Add settings screen and associated navigation for home screens
Many of the changes here are renaming Screens code to be OnboardingScreens.
There's then the addition of HomeScreens as we now have Home as well as
Settings. A navigation controller was then added so that settings can be
navigated to, and some initial settings values added. These values end
up in SharedPreferences where they can be accessed by other code.