Move to using androidx.preference.PreferenceManager
A client can now use code like this to get values set by the settings:

  val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this)
  val allowCallouts = sharedPreferences.getBoolean("allow_callouts", true)

Until the settings page has a value changed, the getBoolean call will
return the default value (true) in this case. The default value has to
be the same as in the Settings Composable code in order not to cause
confusion.