Start implementing LocationDetails screen
This screen is widely used in the iOS app. It can be used to show the
details of a searched location, the current location or a saved
marker. It has a map and a text description which varies depending
on what type of location it is. The map opens at a fixed zoom around
a marker for the location. It also should show the crrent location if
that is within the bounds of the map.

This is a very rough screen which uses Ramani Maps which is a composable
wrapper for MapLibre. The string used to navigate to the LocationDetails
screen is appended with JSON containing the data required by the screen.
That JSON is parsed when the LocationDetails navigation takes place.