Add clearing of the speech queue to the audio engine
The default behaviour of createTextToSpeech is to queue the speech and
play it out once any currently playing speech has finished. However,
some of the UI interactions need to be able to cancel the current
speech and start afresh. This change adds clearTextToSpeechQueue which
does just this.

In the C++ code there's some more care taken so that only currently
playing audio ends up in m_Beacons and all text to speech queued
audio is in m_QueuedBeacons. This means that ClearQueue on the C++
side just has to delete all of the PositionedAudio objects in
m_QueuedBeacons to clear the queue. However, the TextToSpeech engine
in Kotlin needs to take action to close the sockets for the queued
speech or it will just block waiting for the data to be read. We
already had code to do this on destroying the AudioEngine, so this
is now down within the new clearTextToSpeechQueue function too.
Forcing the ending of synthesis required some extra care with the
UtteranceProgressListener.