Initial Route import and playback
This is for some very initial testing and development work. If the user
selects "Open with Soundscape" on a GPX file from the Android File app
then the GPX will be parsed and an inMemory (temporary) database will be
created. There's then a new RoutePlayer class which is created to playback
the first route that can be found in that database. There's some speech
describing when the marker changes and the map reflects the location of
the current beacon (marker in the Route).

If the app is launched in any other way, it will work as before with no
route available or played back.