Fix moving navigation symbol to follow center of map
The main fix here is to update the symbol showing the phone location so
that it stays at the center of the map. This uses addOnCameraMoveListener
to track the camera position. The rotate gesture has been disabled as the
rotation is controlled by the compass heading.

The drawable used to indicate the user location has also been changed so
that it's an SVG image, replacing the poor resolution PNG.