Add icon to turn off Street Preview mode
The full blown street preview mode of the iOS Soundscape allows navigating
the remote location. Our current location provider is static and entered
via the LocationDetails screen. This change adds an icon on the Home screen
to show that the Street Preview mode is active and clicking on it will exit
it.
When the location provider changes the HomeViewModel collecting of the
location and orientation has to be restarted with the new providers.