Don't return to LocationDetails screen on rotate
When an intent is used to launch the app that puts it in the
LocationDetails screen it should only navigate there when the app
first starts. Prior to this change it was also opening that screen
whenever the screen is rotated.
There are two changes:

1. Only parse the Intent the first time the service connection is
bound, and not subsequent times.
2. Reset the `Navigator` destination to "" once the LocationDetail
screen has been displayed and then check for that after a rotate.