Throttle direction heading rate from AndroidDirectionProvider
Although the code requests a refresh rate of 50Hz, on my phone at least
updates are received at up to 200Hz. This change keeps the update rate
to the audio engine where the performance is crucial, but limits updates
to flow listeners to 50Hz and only sends the heading when it changes by
more than 1 degree. This greatly reduces the number of redraws that the
map UI does.