Remove Ramani maps from app and reuse MapContainerLibre
MapContainerLibre is making more sense now, so re-using it for the
LocationDetails screen is a better idea than using Ramani maps. This
change makes MapContainerLibre slightly more configurable so that it
can be centered on any point. This allows it to be easily reused.

The map center location and rotation are now independent from the user
location and rotation so that on the Home screen the map can rotate
with the user and on the LocationDetails screen the user icon rotates
with the user and the map stays with the top at North.

There's likely more to do here, but I think this is a useful first
step.