Switch to planetiler vector tiles for test and start intersection detection
The vector tiles generated by planetiler use different layer names (and include
slightly different data) from those served up by the protomaps demo map. This
change set switches all of the test content for the VectorTilesTest code to be
our own generated set of tiles. The only non-default options used were:

  --maxzoom=15 --render_maxzoom=15 --simplify-tolerance-at-max-zoom=0

We're skipping a whole lot of the layers when turning into GeoJSON and only
handling the place, poi and transportation_name layers. What this means is
that the intersection code is actually spotting all of the intersections between
named roads and generating a point in the GeoJSON for them. However, there are
still other issues:

 1. The transportation_name layer is only for named highways/footpaths etc. All
  of the unnamed ones are in the transportation layer (I don't yet understand
  why this distinction is made!).
 2. If a line changes from a path to a road on the map then it will be described
  as an intersection. A good example of this is Hillhead Street in Milngavie at
  -4.3158,55.9421 which turns pedestrianised.
 3. I've currently just made the intersection name a concatenation of the street
  names involved. Unique ids are not making it into the GeoJSON

Otherwise I'm optimistic about this working. To address item 1 either we change
the map generated by planetiler, or we merge the transportation and
transportation_name layers. Named roads currently appear in BOTH layers which I
think is 'daft' so perhaps makinf the changes within planetiler would make most
sense as it would shrink the size of the output. We may then be able to add in
intersections there (which would increase the size of the output) possibly
without then having to rely on the simplify-tolerance-at-max-zoom flag.