Intersection finessing for VectorTileTest
I think this is as far as I'm going to go with detecting intersections from vector tile
data. The changes appear to successfully de-duplicate the data from transportation and
transportation_name and the output has points at every intersection of interest that I
looked at including entering tunnels and where steps are. However, I'm missing some
data to allow roads that make up the intersection to be queried by Adam's code. As a
proof of concept it's okay, though it's hard to be 100% confident in the results and
the intersection data is in a wholly different format from what Adam's current code
would expect.