Fix memory leaks in MapContainerLibre
The listeners and symbol management that happens in the LaunchedEffect
code were not being removed/deleted when the composable was being disposed
of. I tried changing the LaunchedEffect code to be DisposableEffect, but
the MapView had already been destroyed by the time the onDispose code was
called. Instead the clean up code is passed in to the MapViewHelper so that
it is called just before onDestroy is called on the MapView. This resolves
the memory leaks that Leak Canary was spotting when locking/unlocking the
phone or swiping away the app and reopening it with the service still
running.
To ease this behaviour, the call to rememberMapViewWithLifecycle was moved
inside MapContainerLibre and I've confirmed that rendering is still correct.
I've also changed the function signature on onMapLongClick so that it
returns a Boolean. This means that it can just be passed in to the mapLibre
calls without using the arcane syntax that it had previously.