Fix memory leak when stopping Soundscape service
Leak Canary found a memory leak in the LocalBinder class within the
Soundscape service. The implementation was straight from the Android
docs, but I guess it's not so common to disconnect from a service
when the app isn't shutting down too?
The main fix is to move the LocalBinder class out and add to it a
reset() function which removes the reference to the SoundscapeService.
When the service is destroyed it calls reset() on the binder and then
the garbage collection can continue unhindered.