Add support for Media Transport remote controls - Issue #141
In order to receive Media Transport button events in the service we need
to have the service derive from MediaSessionService. We pass into that
a dummy media player based on SimpleBasePlayer and then install a callback
to get the Media button events.
I can see all the events that I've sent from my headphones and remote
control being logged. I've only hooked up skip-next as it's the one that
we already have code for `myLocation`. The others will be simple to hook
up once we've done more of the callout code.