Update settings with some more audio control and refactor
The main aim of this change was to add a list of supported TextToSpeech voices
and control of the rate of the speech. To do this, the Settings screen needed
some work:

 1. It didn't scroll when there were too many settings to fit on screen
 2. It didn't deal with the delay in the TextToSpeech engine initialization

The change fixes both of these. The second is done with a flow from the audio
engine to indicate when the TextToSpeech engine has been configured. When that's
received the Settings screen will recompose with the updated values.

The main issue with the voice names is that their poor! There doesn't seem to
be any way to get a description of the Voice from the API, so we're left with
their names which are ugly to read. There are various online threads which
suggest just mapping these to made up names - or Voice 1/2 etc. but that's not
been done yet.

The second issue is that as with all of the other settings, the service has to
be restarted to use any changes in value.