Update audio engine immediately if any settings change
The main aim of this change is to update the audio engine whenever any
of the audio settings are changed. This is done by listening to changes
on the SharedPreferences from within the NativeAudioEngine. As well as
updating the audio engine, changes to the speech settings also trigger
some speech so that it's easy to fiddle with settings and hear their
results. This behaviour is disabled for AudioEngine created by Hilt
injection as those are running during onboarding when not all of the
preferences have been configured.

Another complication is that prior to this change the SettingsViewModel
was creating its own audio engine just to get the possible
configuration options. This has been changed so that now it connects
to the Soundscape service and uses its audio engine.