If no TTS voices are found allow mis-matched locales
I installed the Vocalizer TTS app which gives me access to a much wider
range of voices. They sound better than Google, are well named, and seem
to do a better job with Street vs. Saint abbreviations. However, the
locale for the voice I chose was "eng" rather than than the "en" which
was in our locale and so the voice was ignored. If no voices are found
when checking the list from the TTS engine this change means that it's
re-checked this time without matching on the locale language.