Move FMOD libraries to jniLibs directory and remove unused versions
There was a problem building on one particular Mac when the FMOD library
wasn't being installed in the APK. This change moves the libraries to
the jniLibs directory which forces them to be installed. We weren't
using the logging version of the FMOD library, so those versions have
been deleted. The CMakeLists.txt file has been updated to handle the
new location.