Switch to using our own protomaps server and built in maplibre style
The main change here is to store as assets the files required to implement
the osm-bright-gl-style in the maplibre UI. These consist of fonts, sprites
and the main style description. The tile server URL used provides the TileJSON
description of the server which includes attribution for the data sources used
in the maps themselves.