Add ability to switch to using protomaps MVT as source of GeoJSON
Changing SOUNDSCAPE_TILE_BACKEND to be false will switch the app to use
the protomaps tiles instead of those from soundscape-backend.

The MVT to GeoJSON test code has been moved out of the unit test and
into it's own file. It's been adjusted to give a more accurate mapping
to what soundscape-backend providers including adding osm_ids based on
the protomap id values. With these in place the various call outs seem
to work, though there's obviously much more to do.

During the process I refactored getGridFeatureCollections to be a much
more concise function and added some enums to share with the calling
functions.

Our VectorTile class has been renamed Tile to match TileGrid because
there's another VectorTile class which is being automatically generated
from the MVT protobuf file. It's confusing to have two classes with the
same name, and it's not possible to easily change the protobuf file, so
I renamed our one.

For now, because the MVT to GeoJSON is a work in progress, for now the
tile database is always discarded when the app is started. That's because
it contains GeoJSON generated by the conversion code and we want it to
update when the code changes. We'll have to think of a longer term
solution to this (versioning or the protobufs go in the tile database and
they are always reparsed? Or we just rely on the HTTP cache?).