Adjust vectorTileToGeoJson to not use transportation_name layer
We've added names to all lines that have a neam in the transportation
layer and so we no longer need to create a dictionary from the
transportation_name layer. We do this because that layer had merged lines
which lost OSM ids and made de-duplication impossible. The names are only
stored once in the MVT so it's a fairly efficient addition.

I've also added some more test tiles to increase testing coverage.