Allow passing in of zoom level to vectorTileToGeoJson
This is so that tests can pass in the zoom level without relying on it
being set correctly in ZOOM_LEVEL. This allows soundscape-backend and
protomaps backend both to be tested by unit tests.

Add a check to vectorTileToGeoJsonFromFile to ensure that all of the
points within the featureCollection are within the extents of the
buffered tile.