Add cropping of roads and paths to within the tile boundaries
Interpolated points are added to lines which cross the tile boundary
so that they can be accurately cropped. If a line comes back on to
the tile, then a new segment is created again interpolating to the
point at which it crosses back into the tile.
The accuracy is such that if I view the GeoJSON using JOSM the
interpolated points from two adjacent tiles appear to be only a few
centimetres apart. This is better than the large overlaps which we
previously had and should make it easier to join up between tiles.