Add road/path joining at tile boundaries to getGridFeatureCollections()
getGridFeatureCollections is called to get the current tile grid whenever
a callout is about to be calculated. This change adds tiny roads to join
any roads/paths which cross the tile boundaries.

As part of this, the tile database schema changes. I've added code so that
if there's an exception opening the database is simply deletes it and tries
again with a fresh database. That's okay for now, though may need more
thinking about in future.