Remove database for tiles
Instead of using the database as a tile cache or decoded GeoJSON, simply do
all of the conversion work whenever the tile grid changes and keep the result
in memory. getGridFeatureCollections returns the result instantly from memory
and so callouts have less delay.
The implementation doesn't try to be smart so it throws away the old grid and
creates the FeatureCollections from scratch each time. It doesn't reuse tile
data from the tiles that might overlap with the previous grid.
The change works for both soundscape-backend and the protomaps backend. The
time taken to create the grid is considerably longer for soundscape-backend
which obviously might explain why it was originally required (~400ms vs.
~3 seconds for the whole grid).