Add function to get 2x2 grid from MVT files for unit tests
In order to test the MVT -> GeoJSON -> intersection code in total we want
to process the MVT files within the IntersectionsTestMvt tests. This change
adds a helper function getGeoJsonForLocation which can be used by those tests
to create a FeatureCollection from the MVT resources files. update.sh has
been updated to fetch the tiles required for those tests.