Fully populate possible highway types from openmaptiles
The previous list was only partial. This change adds all the types listed
in the openmaptiles mapping.yaml for the transportation layer. The protomaps
layer is condensing some road types (`residential`, `living_street`, `road`)
into a single type of `minor`. It also does this for paths, though the subclass
is used to store the data that was condensed e.g. `pedestrian`, `footway` etc.