Add entrance features including the name of the POI
The code here is similar to the road intersection detection code in that it's
matching up coincedent point in a tile. In this case finding the POI polygon
for an entrance. At present it still uses the OSM id for the entrance rather
than the POI, but it gives us the possiblity of extending the behaviour in
future.