Alter the pixel ratio used when rendering the map
On my phone at least, the text on maps is too small and the roads too
narrow to easily read. Zooming in doesn't help with this as they are
still small, just spread further apart. Changing the pixelRatio on
the map to be 4.0 results in larger text, icons and roads which is
better size wise. The rendering is a bit 'soft', presumably because
it's being rendered at 1/16th size and then scaled up, but it's still
an overall win.