Use rtree to speed up geo searches
The rtree library provides a fast spatial search implementation. There's
a good introduction here:

https://blog.mapbox.com/a-dive-into-spatial-search-algorithms-ebd0c5e39d2a

On my phone the MvtPerformanceTest reduces a search for POI within a certain
distance from ~85ms to ~2ms. There's an initial cost of creating the Rtree
of ~500ms at the point at which the MVT is loaded. Auto callouts are
constantly searching for nearby POI and so this should be an overall gain.

Lines and rectangles can also be included in the rtree, though not arbitrary
polygons. Because we're using longitude and latitude we have to provide our
own set of functions for determining the distance between a point and another
point, a point and a line, and a point and a rectangle. When creating the
rtree I'm entering a bounding box rectangle in place of any polygons. Possibly
entering a polygon as a number or lines might be better? It wouldn't tell us
that we were inside the polygon, but it would perhaps give us a hint?

The current integration of the rtrees is that they are now used in place of
the FeatureCollections inside getGridFeatureCollection. The code now creates
dynamically selected FeatureCollections based on the location and distance
parameters passed in. This allows callers to reduce the number of Features
that they have to process from 'all those in the grid' to 'those within 100'.