Add intersection auto callouts
iOS has an event bus which prioritizes different types of callout. It's not really
clear why they don't just do them all at once. This is my first attempt at calling
out intersections. It uses code from the AheadOfMe. It's currently hard to test
because the simulator is always pointing north and so doesn't always point towards
the junction that is being approached.