Fix up intersection code to deal with more local roads selection
getGridFeatureCollection now returns a small local subset of roads near
to the current location. The result is that it's possible for
roadsGridFeatureCollection to be empty. In that case getNearestRoad was
returning an empty Feature which added to the confusion. These changes
check for an empty collection, and also getNearestRoad returns an
empty collection if that's what's passed in to it.