Add script to convert from iOS string resources to Android and convert all
The instructions in the script are all based on my folder config, but with
a little tweaking will work for anyone. However, we should only need to do
this one time as the iOS strings aren't changing. All new strings have been
manually added at the end of the XML files after the iOS ones.

The string resources within the XML are now surrounded by double quotes.
This removes the need to escape the apostrophes which appear in many strings.

I've also added in the extra regions supported by iOS - two regions of
Portugese, two of English and two of French. However, I haven't managed to
make the app use the regional resources.