Split Location and Direction Providers into their own files
All of the providers were hidden inside LocationProvider, this change
splits them out. It also adds GpxDrivenProvider which is the start of
a combined provider that is driven by GPX files. And finally it adds
updating of the current location to StaticLocationProvider. This allows
the app to move the location for StreetPreview purposes.