Creating a FeatureTreeCollection for Roads resulted in many duplicates
The way that we add lines to the RTree is that each line segment is added
pointing to the same feature. However, this means that a search will return
all the RTree entries which are within the search criteria which means the
same feature will be duplicated depending on how many of its segments fall
within the search area. This change de-duplicates the feature addition so
that each feature is only added a single time.