StreetPreview improvements for handling non-intersections
Several changes in functionality here:

1. The initial StreetPreview jump is now to the nearest node on the
nearest road instead of the nearest intersection.
2. As a result the StreetPreview code has been updated to be able to
split a road at its current location so that it can move either way
along a road.
3. When the StreetPreview hits the edge of the grid, it has to stop
and then allow time to load in the new grid. From that point, it has
to move along the short new joining road which is slightly tricky
because it's pointing in the wrong direction - it will always be
along the grid edge. This commit deals with that by removing the
joining line segment from choice and replacing it with a segment
starting at the current location.

The result is that StreetPreview now seems fairly robust - though if
you jump to a tile grid edge there will be a delay until the next
jump can be made.