Remove trimming of resources to shortened list of languages
Resources that aren't used get trimmed during the build process, but this
setting removes all resources that aren't in the list. This includes those
in AppCompat and Google Play Services libraries which the app uses. Given
that those strings might be used, I think it's reasonable to keep them in
and it takes away another place where we have to list all of the languages
that we support and the risk of it getting out of sync.