Change approach to thread safety of FeatureTrees
The FeatureTrees are updated when we the tile grid is updated. Because
the FeatureTrees are also being used in all of the callout and street
preview calculations we need to take care when we access them. This
change puts all access in a single threaded context - treeContext -
which means that the assigning of the trees to new data, and all of
the callout and street preview calculations should be safely using
the FeatureTrees.