Prevent drawing behind status bar
When an app is running on Android 15 edge-to-edge mode is enabled
by default. As a result the Settings and LocationDetails screens on
my Pixel 8 are currently drawing behind the system status bar. This
change fixes that - though I'm not sure on what the convention is for
setting and passing Modifier?