Playouts all auto callouts in one
When in StreetPreview mode, the location only updates at intersections and not in
between. That meant that only intersections were being called out and not POIs as
the autocallout would only do the first one it found. This change means that it
build all of the callouts that it can on the first change in location.