Add FOV (field of view) searching to FeatureTree class
Prior to this change we could get FeatureCollections by distance from
a point, limiting the results returned to a maximum number if we desired.
This change adds:

generateNearestFeatureCollectionWithinTriangle - returns a FeatureCollection
of the nearest features within the provided FOV triangle.

generateFeatureCollectionWithinTriangle - returns a FeatureCollection
of all the features within the provided FOV triangle.

getNearestFeatureWithinTriangle - returns the nearest Feature within the
FOV triangle.