Refactor some TileUtils functionality
The aim is to simplify the calling code so that commonly used code is hidden
away in shared functions:

* Change functions which either take FeatureCollections containing a single
Feature as arguments or return them have been changed to take/return Feature?
This makes it clearer what the behaviour is. Updating all of the callers then
makes it more obvious whether the FeatureCollection behaviour is required at
all.

* Switch to using the FeatureTree FOV functions which will search within the
passed in FOV triangle. The RTree does very fast filtering based on the
bounding rectangle of the FOV and then runs more fine grained searching on the
much smaller number of filtered Features.

* Start to move away from using `distance_to` within Feature. By using RTree
we are massively shrinking the number of Features being handled and
recalculating the distance/bounding box is low overhead for a few Features.

* Move away from functions with Poi/Intersection/Crossing in the title where
the function deals with generic Features.

* The unit tests were all written to use FeatureCollections directly, and
rather than change them too much, I switch to using the FeatureTree versions
of functions by creating temporary FeatureTrees each time they were required.
This isn't how the main application behaves, but there's a lot of duplicated code
in the tests that could do with centralizing (in the tests at least).