Centralize intersection description code
The main change here is to de-duplicate the intersection description code
from the unit tests and GeoEngine and have a single function in a new file
IntersectionUtils. This should be the same as as that which was in the
ComplexIntersections unit test. A new function in the same file can then
take the IntersectionDescription and turn it into a callout.
The other change in that area is the start of improving the callout code
to use the results of the ActivityRecognition to switch between heading,
facing and travelling in the callouts.
The final change is that getNearestRoad now uses FeatureTree to perform
it's searching. Not only does this improve its performance, but it allows
us to search outside the FOV - the nearest road could be 0.5m behind the
current FOV and searching only within the FOV would preclude that. The
unit tests all still pass with this change.